feat: update services and handlers to include accountID parameter for transaction methods

+++ internal/utils/auth_context.go
@@ -7,7 +7,7 @@ import (
 )
 
 // GetUserIDFromLocals parses user/account ID from Fiber locals.
-func GetUserIDFromLocals(c *fiber.Ctx) (int64, bool) {
+func GetUserIDFromLocals(c *fiber.Ctx) (uint, bool) {
 	userID := c.Locals("userID")
 	if userID == nil {
 		userID = c.Locals("accountID")
@@ -15,19 +15,19 @@ func GetUserIDFromLocals(c *fiber.Ctx) (int64, bool) {
 
 	switch value := userID.(type) {
 	case int64:
-		return value, value > 0
+		return uint(value), value > 0
 	case uint:
-		return int64(value), value > 0
+		return value, value > 0
 	case int:
-		return int64(value), value > 0
+		return uint(value), value > 0
 	case float64:
-		return int64(value), value > 0
+		return uint(value), value > 0
 	case string:
 		parsed, err := strconv.ParseInt(value, 10, 64)
 		if err != nil || parsed <= 0 {
 			return 0, false
 		}
-		return parsed, true
+		return uint(parsed), true
 	default:
 		return 0, false
 	}
